module06/theme02/seminar/chatbotbackend/src/schemes.py
from openai.types.chat import (
    ChatCompletionAssistantMessageParam,
    ChatCompletionUserMessageParam,
)
from pydantic import BaseModel


class ChatRequest(BaseModel):
    query: str
    history: (
        list[ChatCompletionUserMessageParam | ChatCompletionAssistantMessageParam]
        | None
    ) = None
