module06/theme02/seminar/chatbotbackend/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.123.7",
    "openai>=2.8.1",
    "pydantic-settings>=2.12.0",
    "qdrant-client>=1.16.1",
]

[dependency-groups]
lint = [
    "pre-commit>=4.5.0",
    "ruff>=0.14.7",
]
type = [
    "mypy>=1.19.0",
]


[tool.ruff]
target-version = "py310"


[tool.ruff.lint]
select = [
    "F",       # pyflakes rules
    "I",       # sorting for imports
    "UP",      # upgrade to latest syntax if possible
    "C4",      # cleaner comprehensions
    "NPY",     # numpy
    "N",       # PEP8 naming convention
    "RUF009",  # function-call-in-dataclass-default-argument
    "RUF008",  # mutable-dataclass-default
    "RUF013",  # implicit-optional
    "RUF012",  # mutable-class-default
    "RUF015",  # unnecessary-iterable-allocation-for-first-element
    "RUF016",  # invalid-index-type
    "RUF017",  # quadratic-list-summation
    "RUF018",  # assignment-in-assert
    "RUF019",  # unnecessary-key-check
    "RUF020",  # never-union
    "RUF021",  # parenthesize-chained-operators
    "RUF022",  # unsorted-dunder-all
    "RUF024",  # mutable-fromkeys-value
    "RUF026",  # default-factory-kwarg
    "RUF033",  # post-init-default
    "RUF034",  # useless-if-else
    "RUF041",  # unnecessary-nested-literal
    "RUF040",  # invalid-assert-message-literal-argument
    "RUF046",  # unnecessary-cast-to-int
    "RUF051",  # if-key-in-dict-del
    "RUF100",  # unused-noqa
    "RUF101",  # redirected-noqa
    "RUF200",  # invalid-pyproject-toml
    "PTH",     # use pathlib
    "TID",     # tidy-imports
    "FAST",    # fastapi-specific rules
]

[tool.mypy]
plugins = ['pydantic.mypy']